Refactored readonly code
